+++ crates/matrix-sdk-ui/src/timeline/controller/mod.rs
@@ -70,8 +70,9 @@ use super::{
     subscriber::TimelineSubscriber,
     traits::{Decryptor, RoomDataProvider},
     DateDividerMode, Error, EventSendState, EventTimelineItem, InReplyToDetails, Message,
-    PaginationError, Profile, RepliedToEvent, TimelineDetails, TimelineEventItemId, TimelineFocus,
-    TimelineItem, TimelineItemContent, TimelineItemKind,
+    PaginationError, Profile, ReactionsByKeyBySender, RepliedToEvent, TimelineDetails,
+    TimelineEventItemId, TimelineFocus, TimelineItem, TimelineItemContent, TimelineItemKind,
+    VirtualTimelineItem,
 };
 use crate::{
     timeline::{
@@ -79,7 +80,7 @@ use crate::{
         date_dividers::DateDividerAdjuster,
         event_item::EventTimelineItemKind,
         pinned_events_loader::{PinnedEventsLoader, PinnedEventsLoaderError},
-        TimelineEventFilterFn,
+        AggregatedTimelineItemContent, AggregatedTimelineItemContentKind, TimelineEventFilterFn,
     },
     unable_to_decrypt_hook::UtdHookManager,
 };
@@ -1027,12 +1028,15 @@ impl<P: RoomDataProvider, D: Decryptor> TimelineController<P, D> {
 
         // Replace the local-related state (kind) and the content state.
         let prev_reactions = prev_item.content().reactions();
+        let prev_thread_root = prev_item.content().thread_root();
+        let prev_in_reply_to = prev_item.content().in_reply_to();
         let new_item = TimelineItem::new(
             prev_item.with_kind(ti_kind).with_content(TimelineItemContent::message(
                 content,
                 None,
-                &txn.items,
                 prev_reactions,
+                prev_thread_root,
+                prev_in_reply_to,
             )),
             prev_item.internal_id.to_owned(),
         );
@@ -1309,6 +1313,18 @@ impl<P: RoomDataProvider, D: Decryptor> TimelineController<P, D> {
             }
         }
     }
+
+    /// Insert a timeline start item at the beginning of the room, if it's
+    /// missing.
+    pub async fn insert_timeline_start_if_missing(&self) {
+        let mut state = self.state.write().await;
+        let mut txn = state.transaction();
+        if txn.items.get(0).is_some_and(|item| item.is_timeline_start()) {
+            return;
+        }
+        txn.items.push_front(txn.meta.new_timeline_item(VirtualTimelineItem::TimelineStart), None);
+        txn.commit();
+    }
 }
 
 impl TimelineController {
@@ -1328,11 +1344,17 @@ impl TimelineController {
             .ok_or(Error::EventNotInTimeline(TimelineEventItemId::EventId(event_id.to_owned())))?
             .clone();
 
-        let TimelineItemContent::Message(message) = item.content().clone() else {
+        let TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message),
+            reactions,
+            thread_root,
+            in_reply_to,
+        }) = item.content().clone()
+        else {
             debug!("Event is not a message");
             return Ok(());
         };
-        let Some(in_reply_to) = message.in_reply_to() else {
+        let Some(in_reply_to) = in_reply_to else {
             debug!("Event is not a reply");
             return Ok(());
         };
@@ -1353,6 +1375,8 @@ impl TimelineController {
             &item,
             internal_id,
             &message,
+            &reactions,
+            &thread_root,
             &in_reply_to.event_id,
             self.room(),
         )
@@ -1366,11 +1390,17 @@ impl TimelineController {
 
         // Check the state of the event again, it might have been redacted while
         // the request was in-flight.
-        let TimelineItemContent::Message(message) = item.content().clone() else {
+        let TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message),
+            reactions,
+            thread_root,
+            in_reply_to,
+        }) = item.content().clone()
+        else {
             info!("Event is no longer a message (redacted?)");
             return Ok(());
         };
-        let Some(in_reply_to) = message.in_reply_to() else {
+        let Some(in_reply_to) = in_reply_to else {
             warn!("Event no longer has a reply (bug?)");
             return Ok(());
         };
@@ -1380,12 +1410,12 @@ impl TimelineController {
         trace!("Updating in-reply-to details");
         let internal_id = item.internal_id.to_owned();
         let mut item = item.clone();
-        item.set_content(TimelineItemContent::Message(
-            message.with_in_reply_to(InReplyToDetails {
-                event_id: in_reply_to.event_id.clone(),
-                event,
-            }),
-        ));
+        item.set_content(TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message),
+            reactions,
+            thread_root,
+            in_reply_to: Some(InReplyToDetails { event_id: in_reply_to.event_id, event }),
+        }));
         state.items.replace(index, TimelineItem::new(item, internal_id));
 
         Ok(())
@@ -1493,12 +1523,15 @@ impl<P: RoomDataProvider> TimelineController<P, (OlmMachine, OwnedRoomId)> {
     }
 }
 
+#[allow(clippy::too_many_arguments)]
 async fn fetch_replied_to_event(
     mut state: RwLockWriteGuard<'_, TimelineState>,
     index: usize,
     item: &EventTimelineItem,
     internal_id: TimelineUniqueId,
     message: &Message,
+    reactions: &ReactionsByKeyBySender,
+    thread_root: &Option<OwnedEventId>,
     in_reply_to: &EventId,
     room: &Room,
 ) -> Result<TimelineDetails<Box<RepliedToEvent>>, Error> {
@@ -1511,11 +1544,18 @@ async fn fetch_replied_to_event(
     // Replace the item with a new timeline item that has the fetching status of the
     // replied-to event to pending.
     trace!("Setting in-reply-to details to pending");
-    let reply = message.with_in_reply_to(InReplyToDetails {
+    let in_reply_to_details = Some(InReplyToDetails {
         event_id: in_reply_to.to_owned(),
         event: TimelineDetails::Pending,
     });
-    let event_item = item.with_content(TimelineItemContent::Message(reply));
+
+    let event_item =
+        item.with_content(TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message.clone()),
+            reactions: reactions.clone(),
+            thread_root: thread_root.clone(),
+            in_reply_to: in_reply_to_details,
+        }));
 
     let new_timeline_item = TimelineItem::new(event_item, internal_id);
     state.items.replace(index, new_timeline_item);