+++ crates/matrix-sdk-ui/src/timeline/tests/edit.rs
@@ -32,7 +32,9 @@ use ruma::{
 use stream_assert::{assert_next_matches, assert_pending};
 
 use super::TestTimeline;
-use crate::timeline::TimelineItemContent;
+use crate::timeline::{
+    AggregatedTimelineItemContent, AggregatedTimelineItemContentKind, TimelineItemContent,
+};
 
 #[async_test]
 async fn test_live_redacted() {
@@ -74,7 +76,12 @@ async fn test_live_sanitized() {
 
     let item = assert_next_matches!(stream, VectorDiff::PushBack { value } => value);
     let first_event = item.as_event().unwrap();
-    assert_let!(TimelineItemContent::Message(message) = first_event.content());
+    assert_let!(
+        TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message),
+            ..
+        }) = first_event.content()
+    );
     assert_let!(MessageType::Text(text) = message.msgtype());
     assert_eq!(text.body, "**original** message");
     assert_eq!(text.formatted.as_ref().unwrap().body, "<strong>original</strong> message");
@@ -99,7 +106,12 @@ async fn test_live_sanitized() {
 
     let item = assert_next_matches!(stream, VectorDiff::Set { index: 1, value } => value);
     let first_event = item.as_event().unwrap();
-    assert_let!(TimelineItemContent::Message(message) = first_event.content());
+    assert_let!(
+        TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message),
+            ..
+        }) = first_event.content()
+    );
     assert_let!(MessageType::Text(text) = message.msgtype());
     assert_eq!(text.body, new_plain_content);
     assert_eq!(text.formatted.as_ref().unwrap().body, " <strong>better</strong> message");
@@ -144,7 +156,12 @@ async fn test_aggregated_sanitized() {
 
     let item = assert_next_matches!(stream, VectorDiff::PushBack { value } => value);
     let first_event = item.as_event().unwrap();
-    assert_let!(TimelineItemContent::Message(message) = first_event.content());
+    assert_let!(
+        TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message),
+            ..
+        }) = first_event.content()
+    );
     assert_let!(MessageType::Text(text) = message.msgtype());
     assert_eq!(text.body, "!!edited!! **better** message");
     assert_eq!(text.formatted.as_ref().unwrap().body, " <strong>better</strong> message");
@@ -196,7 +213,12 @@ async fn test_edit_updates_encryption_info() {
         VerificationState::Verified
     );
 
-    assert_let!(TimelineItemContent::Message(message) = first_event.content());
+    assert_let!(
+        TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message),
+            ..
+        }) = first_event.content()
+    );
     assert_let!(MessageType::Text(text) = message.msgtype());
     assert_eq!(text.body, "**original** message");
 
@@ -225,7 +247,12 @@ async fn test_edit_updates_encryption_info() {
         VerificationState::Unverified(VerificationLevel::UnverifiedIdentity)
     );
 
-    assert_let!(TimelineItemContent::Message(message) = first_event.content());
+    assert_let!(
+        TimelineItemContent::Aggregated(AggregatedTimelineItemContent {
+            kind: AggregatedTimelineItemContentKind::Message(message),
+            ..
+        }) = first_event.content()
+    );
     assert_let!(MessageType::Text(text) = message.msgtype());
     assert_eq!(text.body, "!!edited!! **better** message");
 }