+++ crates/matrix-sdk/src/authentication/oauth/account_management_url.rs
@@ -16,8 +16,10 @@
 //!
 //! This is a Matrix extension introduced in [MSC4191](https://github.com/matrix-org/matrix-spec-proposals/pull/4191).
 
-use ruma::OwnedDeviceId;
-use serde::Serialize;
+use ruma::{
+    api::client::discovery::get_authorization_server_metadata::msc2965::AccountManagementAction,
+    OwnedDeviceId,
+};
 use url::Url;
 
 /// An account management action that a user can take, including a device ID for
@@ -26,26 +28,22 @@ use url::Url;
 /// The actions are defined in [MSC4191].
 ///
 /// [MSC4191]: https://github.com/matrix-org/matrix-spec-proposals/pull/4191
-#[derive(Debug, Clone, PartialEq, Eq, Serialize)]
-#[serde(tag = "action")]
+#[derive(Debug, Clone, PartialEq, Eq)]
 #[non_exhaustive]
 pub enum AccountManagementActionFull {
     /// `org.matrix.profile`
     ///
     /// The user wishes to view their profile (name, avatar, contact details).
-    #[serde(rename = "org.matrix.profile")]
     Profile,
 
     /// `org.matrix.sessions_list`
     ///
     /// The user wishes to view a list of their sessions.
-    #[serde(rename = "org.matrix.sessions_list")]
     SessionsList,
 
     /// `org.matrix.session_view`
     ///
     /// The user wishes to view the details of a specific session.
-    #[serde(rename = "org.matrix.session_view")]
     SessionView {
         /// The ID of the session to view the details of.
         device_id: OwnedDeviceId,
@@ -54,7 +52,6 @@ pub enum AccountManagementActionFull {
     /// `org.matrix.session_end`
     ///
     /// The user wishes to end/log out of a specific session.
-    #[serde(rename = "org.matrix.session_end")]
     SessionEnd {
         /// The ID of the session to end.
         device_id: OwnedDeviceId,
@@ -63,53 +60,106 @@ pub enum AccountManagementActionFull {
     /// `org.matrix.account_deactivate`
     ///
     /// The user wishes to deactivate their account.
-    #[serde(rename = "org.matrix.account_deactivate")]
     AccountDeactivate,
 
     /// `org.matrix.cross_signing_reset`
     ///
     /// The user wishes to reset their cross-signing keys.
-    #[serde(rename = "org.matrix.cross_signing_reset")]
     CrossSigningReset,
 }
 
-/// Build the URL for accessing the account management capabilities, as defined
-/// in [MSC].
+impl AccountManagementActionFull {
+    /// Get the [`AccountManagementAction`] matching this
+    /// [`AccountManagementActionFull`].
+    pub fn action_type(&self) -> AccountManagementAction {
+        match self {
+            Self::Profile => AccountManagementAction::Profile,
+            Self::SessionsList => AccountManagementAction::SessionsList,
+            Self::SessionView { .. } => AccountManagementAction::SessionView,
+            Self::SessionEnd { .. } => AccountManagementAction::SessionEnd,
+            Self::AccountDeactivate => AccountManagementAction::AccountDeactivate,
+            Self::CrossSigningReset => AccountManagementAction::CrossSigningReset,
+        }
+    }
+
+    /// Append this action to the query of the given URL.
+    fn append_to_url(&self, url: &mut Url) {
+        let mut query_pairs = url.query_pairs_mut();
+        query_pairs.append_pair("action", self.action_type().as_str());
+
+        match self {
+            Self::SessionView { device_id } | Self::SessionEnd { device_id } => {
+                query_pairs.append_pair("device_id", device_id.as_str());
+            }
+            _ => {}
+        }
+    }
+}
+
+/// Builder for the URL for accessing the account management capabilities, as
+/// defined in [MSC4191].
 ///
-/// # Arguments
+/// This type can be instantiated with [`OAuth::account_management_url()`] and
+/// [`OAuth::fetch_account_management_url()`].
 ///
-/// * `account_management_uri` - The URL to access the issuer's account
-///   management capabilities.
+/// [`AccountManagementUrlBuilder::build()`] returns a URL to be opened in a web
+/// browser where the end-user will be able to access the account management
+/// capabilities of the issuer.
 ///
-/// * `action` - The action that the user wishes to take.
+/// # Example
 ///
-/// # Returns
+/// ```no_run
+/// use matrix_sdk::authentication::oauth::AccountManagementActionFull;
+/// # _ = async {
+/// # let client: matrix_sdk::Client = unimplemented!();
+/// let oauth = client.oauth();
 ///
-/// A URL to be opened in a web browser where the end-user will be able to
-/// access the account management capabilities of the issuer.
+/// // Get the account management URL from the server metadata.
+/// let Some(url_builder) = oauth.account_management_url().await? else {
+///     println!("The server doesn't advertise an account management URL");
+///     return Ok(());
+/// };
 ///
-/// # Errors
+/// // The user wants to see the list of sessions.
+/// let url =
+///     url_builder.action(AccountManagementActionFull::SessionsList).build();
 ///
-/// Returns an error if serializing the action fails.
+/// println!("See your sessions at: {url}");
+/// # anyhow::Ok(()) };
+/// ```
 ///
 /// [MSC4191]: https://github.com/matrix-org/matrix-spec-proposals/pull/4191
-pub(crate) fn build_account_management_url(
-    mut account_management_uri: Url,
-    action: AccountManagementActionFull,
-) -> Result<Url, serde_html_form::ser::Error> {
-    let extra_query = serde_html_form::to_string(action)?;
+/// [`OAuth::account_management_url()`]: super::OAuth::account_management_url
+/// [`OAuth::fetch_account_management_url()`]: super::OAuth::fetch_account_management_url
+#[derive(Debug, Clone)]
+pub struct AccountManagementUrlBuilder {
+    account_management_uri: Url,
+    action: Option<AccountManagementActionFull>,
+}
 
-    // Add our parameters to the query, because the URL might already have one.
-    let mut full_query = account_management_uri.query().map(ToOwned::to_owned).unwrap_or_default();
-
-    if !full_query.is_empty() {
-        full_query.push('&');
+impl AccountManagementUrlBuilder {
+    /// Construct an [`AccountManagementUrlBuilder`] for the given URL.
+    pub(super) fn new(account_management_uri: Url) -> Self {
+        Self { account_management_uri, action: None }
     }
-    full_query.push_str(&extra_query);
 
-    account_management_uri.set_query(Some(&full_query));
+    /// Set the action that the user wishes to take.
+    pub fn action(mut self, action: AccountManagementActionFull) -> Self {
+        self.action = Some(action);
+        self
+    }
 
-    Ok(account_management_uri)
+    /// Build the URL to present to the end user.
+    pub fn build(self) -> Url {
+        // Add our parameters to the query, because the URL might already have one.
+        let mut account_management_uri = self.account_management_uri;
+
+        if let Some(action) = &self.action {
+            action.append_to_url(&mut account_management_uri);
+        }
+
+        account_management_uri
+    }
 }
 
 #[cfg(test)]
@@ -117,55 +167,50 @@ mod tests {
     use ruma::owned_device_id;
     use url::Url;
 
-    use super::{build_account_management_url, AccountManagementActionFull};
+    use super::{AccountManagementActionFull, AccountManagementUrlBuilder};
 
     #[test]
     fn test_build_account_management_url_actions() {
         let base_url = Url::parse("https://example.org").unwrap();
         let device_id = owned_device_id!("ABCDEFG");
 
-        let url =
-            build_account_management_url(base_url.clone(), AccountManagementActionFull::Profile)
-                .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone()).build();
+        assert_eq!(url, base_url);
+
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::Profile)
+            .build();
         assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.profile");
 
-        let url = build_account_management_url(
-            base_url.clone(),
-            AccountManagementActionFull::SessionsList,
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::SessionsList)
+            .build();
         assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.sessions_list");
 
-        let url = build_account_management_url(
-            base_url.clone(),
-            AccountManagementActionFull::SessionView { device_id: device_id.clone() },
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::SessionView { device_id: device_id.clone() })
+            .build();
         assert_eq!(
             url.as_str(),
             "https://example.org/?action=org.matrix.session_view&device_id=ABCDEFG"
         );
 
-        let url = build_account_management_url(
-            base_url.clone(),
-            AccountManagementActionFull::SessionEnd { device_id },
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::SessionEnd { device_id })
+            .build();
         assert_eq!(
             url.as_str(),
             "https://example.org/?action=org.matrix.session_end&device_id=ABCDEFG"
         );
 
-        let url = build_account_management_url(
-            base_url.clone(),
-            AccountManagementActionFull::AccountDeactivate,
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::AccountDeactivate)
+            .build();
         assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.account_deactivate");
 
-        let url =
-            build_account_management_url(base_url, AccountManagementActionFull::CrossSigningReset)
-                .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url)
+            .action(AccountManagementActionFull::CrossSigningReset)
+            .build();
         assert_eq!(url.as_str(), "https://example.org/?action=org.matrix.cross_signing_reset");
     }
 
@@ -173,16 +218,16 @@ mod tests {
     fn test_build_account_management_url_with_query() {
         let base_url = Url::parse("https://example.org/?sid=123456").unwrap();
 
-        let url =
-            build_account_management_url(base_url.clone(), AccountManagementActionFull::Profile)
-                .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url.clone())
+            .action(AccountManagementActionFull::Profile)
+            .build();
         assert_eq!(url.as_str(), "https://example.org/?sid=123456&action=org.matrix.profile");
 
-        let url = build_account_management_url(
-            base_url,
-            AccountManagementActionFull::SessionView { device_id: owned_device_id!("ABCDEFG") },
-        )
-        .unwrap();
+        let url = AccountManagementUrlBuilder::new(base_url)
+            .action(AccountManagementActionFull::SessionView {
+                device_id: owned_device_id!("ABCDEFG"),
+            })
+            .build();
         assert_eq!(
             url.as_str(),
             "https://example.org/?sid=123456&action=org.matrix.session_view&device_id=ABCDEFG"